Small Interpreter module refactoring
Extracted types and Utils from the main Interpreter module into submodules.
