+++ src/Interpreter.hs
@@ -3,6 +3,7 @@ module Interpreter
       interpret
   ) where
 
+
 import qualified Data.Map             as M
 
 import           Control.Monad.Except
@@ -11,7 +12,8 @@ import           Control.Monad.State
 
 import qualified Baalbolge.Abs        as BG
 
-import           Baalbolge.Print
+import           Interpreter.Types
+import           Interpreter.Util
 import           Types
 
 
@@ -19,6 +21,8 @@ import           Types
 -}
 interpret :: BG.Exps -> Err Result
 interpret (BG.Program _ exps) = runReader (runExceptT $ interpretExps exps) M.empty
+
+
 
 {- | Interprets a list of expressions and resturns the result value.
 
@@ -42,6 +46,8 @@ interpretExpsRec RUnit (h:t) = do
         rv    -> return rv
 -- | if the type of TUnit or TBool, or the list is empty, we simply return the type
 interpretExpsRec r _ = return r
+
+
 
 {- | Interprets a single expression.
 
@@ -68,9 +74,12 @@ interpretExp (BG.EInternal _ int) = interpretIFunc int
 interpretExp e = throwError $ "Interpretation of exp: " ++ show e
     ++ " is not yet implemented"
 
+
+
 {- | Interprets an internal function usage in Baalbolge.
 -}
 interpretIFunc :: BG.InternalFunc -> InterpreterState
+
 {- | Variable declaration computes the value of the expression and checks if it matches
 the requested type. If there's a match, the variable is created and the value is assigned.
 If the types don't match, runtime exception is thrown.
@@ -83,6 +92,7 @@ interpretIFunc iFunc@(BG.IVarDecl pos t (BG.Var v) e) = do
     if typeEq t ex
         then put (M.insert v ex st) >> return RUnit
         else throwError $ typesError iFunc "variable declaration" pos (pprintType t) (pprintResult ex)
+
 {- | When statement computes the condition first. Then, it checks whether the type is
 bool. If it is, then the computation continues and depending on the value (True/False),
 the given expression is computed or not. If the type of the condition isn't bool, runtime
@@ -96,6 +106,7 @@ interpretIFunc iFunc@(BG.IWhen pos cond e) = do
     case condVal of
         RBool b -> if b then interpretExp e else return RUnit
         _ -> throwError $ typesError iFunc "when statement" pos "bool" (pprintResult condVal)
+
 {- | If statement computes the condition first. Then, it checks whether the type is
 bool. If it is, then the computation continues and depending on the value (True/False),
 one of the expressions is computed. If the type of the condition isn't bool, runtime
@@ -108,6 +119,7 @@ interpretIFunc iFunc@(BG.IIf pos cond e1 e2) = do
     case condVal of
         RBool b -> if b then interpretExp e1 else interpretExp e2
         _ -> throwError $ typesError iFunc "if statement" pos "bool" (pprintResult condVal)
+
 {- | While loop computes the condition first. Then, it checks whether the type is
 bool. If it is, then the computation continues and the given expression is evalueated
 if the condition is True. If the type of the condition isn't bool, runtime exception
@@ -133,6 +145,8 @@ interpretIFunc iFunc@(BG.IWhile pos cond e) = do
 interpretIFunc e = throwError $ "Interpretation of internal function: " ++ show e
     ++ " is not yet implemented"
 
+
+
 {- | Interprets a variable and returns its value. If there's no variable of the given name
 in the state, an error is thrown.
 
@@ -155,6 +169,7 @@ interpretVar (BG.Var v) = do
         Just val -> return val
         Nothing  -> throwError $ "variable '" ++ v ++ "' not found!"
 
+
 {- | Interprets the value of a bool.
 
 $setup
@@ -170,55 +185,3 @@ Right (RBool False)
 interpretBool :: BG.Bool -> InterpreterState
 interpretBool (BG.BTrue _)  = return $ RBool True
 interpretBool (BG.BFalse _) = return $ RBool False
-
-{- | Compares the type of variable and the value we try to assign to it. It's an extension
-of type checking in the TypeChecker as type var can produce runtime type incompatibilities.
--}
-typeEq :: BG.Type -> Result -> Bool
-typeEq (BG.TInt _) (RInt _)   = True
-typeEq (BG.TBool _) (RBool _) = True
-typeEq (BG.TVar _) _          = True
-typeEq (BG.TUnit _) RUnit     = True
-typeEq (BG.TUnit _) _         = False
-typeEq _ _                    = False
-
-{- | Handler of 'variable not found' error, which adds details about the error
-for more readability.
--}
-varNotFoundHandler :: Print a => a -> BG.BNFC'Position -> String -> InterpreterState
-varNotFoundHandler ex pos er = throwError $ varNotFoundError ex er pos
-
--- | Creates a message about types error in the code.
-typesError :: Print a => a -> String -> BG.BNFC'Position -> String -> String -> String
-typesError ex op (Just (line, col)) t1 t2 =
-    "Runtime exception! Types don't match! In operation '" ++ op
-    ++ "' in line " ++ show line ++ ", column " ++ show col ++ ":\n  Expected '"
-    ++ t1 ++ "' but got '" ++ t2 ++ "'!\n    (" ++ printTree ex ++ ")"
-typesError ex op Nothing t1 t2 =
-    "Runtime exception! Types don't match! In operation '" ++ op
-    ++ "' at undetermined position:\n  Expected '" ++ t1
-    ++ "' but got '" ++ t2 ++ "'!\n    (" ++ printTree ex ++ ")"
-
-{- | Extends a simple information regaring 'variable not found' error with information
-position of the erorr and the code itself.
--}
-varNotFoundError :: Print a => a -> String -> BG.BNFC'Position -> String
-varNotFoundError ex er (Just (line, col)) = er ++ "\n  In line " ++ show line
-    ++ ", column " ++ show col ++ ":\n    " ++ printTree ex
-varNotFoundError ex er Nothing = er ++ "\n  At undetermined position:\n    "
-    ++ printTree ex
-
--- | Pretty prints the type used in Baalbolge
-pprintType :: BG.Type -> String
-pprintType (BG.TInt _)  = "int"
-pprintType (BG.TBool _) = "bool"
-pprintType (BG.TUnit _) = "unit"
-pprintType (BG.TVar _)  = "var"
--- TODO: implementation of lists
-pprintType _            = undefined
-
--- | Pretty prints the type of a result of an expression
-pprintResult :: Result -> String
-pprintResult RUnit     = "unit"
-pprintResult (RInt _)  = "int"
-pprintResult (RBool _) = "bool"
