+++ src/Interpreter/Types.hs
@@ -1,0 +1,27 @@
+module Interpreter.Types
+    ( -- * Data types
+      InterpreterReader
+      , InterpreterState
+
+      -- * Data structures
+      , Result (..)
+  ) where
+
+
+import qualified Data.Map             as M
+
+import           Control.Monad.Except
+import           Control.Monad.Reader
+import           Control.Monad.State
+
+import Types
+
+type ExT = ExceptT String
+
+type Name = String
+
+type InterpreterEnv = M.Map Name Result
+type InterpreterSt = InterpreterEnv
+
+type InterpreterReader = ExT (Reader InterpreterEnv) Result
+type InterpreterState = ExT (State InterpreterSt) Result
